refactor(auth): split oauth-launchers/index.ts into client + launchers
index.ts was ~393 lines of implementation wearing a barrel's name. Split
along the real dependency seam:

- oauth-client.ts: the PKCE OAuth client core (createOAuthClient), its
  OAuthClientError, and the shared config/storage types the launchers
  consume.
- launchers.ts: createBrowserOAuthLauncher and createExtensionOAuthLauncher,
  which build on the client.
- index.ts: a true barrel, re-exporting the identical public surface so the
  @epicenter/auth/oauth-launchers subpath and every app consumer are
  unchanged.

Implementations moved verbatim; only relocation and intra-folder imports.
